Updated update_revision.sh to use GitHub
Gerrit-related features were removed. After updating the revision on the local sources file, the script now:
- Checks connection to GitHub
- Finds out if the sources file with the latest revision has been previously pushed
- Commits and pushes to the user's fork

update_revision.sh now only takes as parameters the sources file path and the user's GitHub account. It is assumed that whoever is running the script is capable of pushing to the advance-toolchain fork of the provided user without any password.

Signed-off-by: Denilson Antonio Marques Junior <denimj@linux.vnet.ibm.com>