Created AT 11.0 from next
AT 11.0 was created from the next configset. For gcc and glibc, the
sources file in AT 11.0 was made a symlink to the respective files
in AT next. The AT 11.0 gdb sources file was made a symlink to
AT 10.0 gdb sources file. Also, the Python and binutils sources file
were updated on AT 11.0 to reflect newer versions that were already
on the master branch.

Signed-off-by: Denilson Antonio Marques Junior <denimj@linux.vnet.ibm.com>