Improved code readability & minor improvements
The function send_to_review was separated in two, one specific for Gerrit and another for GitHub. There was too little code in common for those to justify the existence of a function to execute common code. The script now also prints the output of cURL in case the pull request creating in GitHub.

Signed-off-by: Denilson Antonio Marques Junior <denimj@linux.vnet.ibm.com>