Improved remote-build.exp to support custom directories
This script now allows for specification of custom paths
for installing and building AT, through the usage of the
optional parameters --destprefix, --trayprefix and --atdir.
Default behavior has not changed, except for the fact that
it's no longer possible to specify $tray_dir and $destdir
directly in a config file.

Signed-off-by: Denilson Antonio Marques Junior
<denimj@linux.vnet.ibm.com>