Stop update_revision.sh from overwriting PRs
There's a problem that update_revision.sh commits a package revision
update and tries to open a new pull request for it even when one
already exists. This is unwanted behavior and this commit is to fix
this. Now it checks if a PR from the specified GitHub user already
exists, and it doesn't do anything if that is the case.

This commit also aborts script execution in case there are no new
changes to the specified package, which was not happening because
the function update_revision was using return instead of exit.

Signed-off-by: Denilson Antonio Marques Junior
<denimj@linux.vnet.ibm.com>