Add support for environment modules
The Advance Toolchain is a good use-case for environment modules because
to effectively use it requires accurately setting a few independent
environment variables (PATH, MANPATH, INFOPATH).  Using an environment
module can do (and undo) this with single, simple commands:
$ module load at11.0
$ module unload at11.0

This patch, for AT 11 and higher:
- adds a dependency of "environment-modules" for the "-devel" package
- generates the environment module definition file automatically, based on
  the version of AT (including unique names for alpha and beta releases)
- includes the modulefile in the -devel package
- creates a symlink from the canonical system modulefiles directory to the
  AT modulefile, making the module visable with "module avail" and usable
  with "module load/unload/show/whatis"
- cleans up the symlink upon removal of the "-devel" package

AT_VER_REV_INTERNAL was used as the name of the environment module, as it
includes the version, release, and internal qualifier (like "alpha"),
which matches the normal installed path of the Advance Toolchain.
In the DEB-based processing, this variable had been modified to remove
any "." characters, presumably to appease "cron" or a similar external
dependency.  I reverted this modification through all versions, and
adjusted the individual usage to match the historical results where needed.

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>