Add support for environment modules
The Advance Toolchain is a good use-case for environment modules because
to effectively use it requires accurately setting a few independent
environment variables (PATH, MANPATH, INFOPATH).  Using an environment
module can do (and undo) this with single, simple commands:
$ module load at11.0
$ module unload at11.0

This patch, for AT 11 and higher:
- generates the environment module modulefiles automatically, based on
  the version of AT (including unique names for alpha and beta releases)
- generates a modulefile for the -cross environment including aliases
  for all cross-compiler commands
- includes the AT modulefile in the -devel package
- includes the AT and AT-cross modulefiles in the cross-common package
- creates a symlink from the canonical system modulefiles directory to the
  AT modulefiles, making the modules visable with "module avail" and usable
  with "module load/unload/show/whatis"
- cleans up the symlinks upon removal of the packages

AT_VER_REV_INTERNAL was used as the name of the environment module, as it
includes the version, release, and internal qualifier (like "alpha"),
which matches the normal installed path of the Advance Toolchain.
In the DEB-based processing, this variable had been modified to remove
any "." characters, presumably to appease "cron" or a similar external
dependency.  I reverted this modification through all versions, and
adjusted the individual usage to match the historical results where needed.

TARGET was used as the qualifier for the cross-compiler modulefile name,
as it will allow for unique names for each cross-compiler (ppc64 and
ppc64le).

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>