Fixes to the cache manager mechanism
This patch changes to 644 the mode of the cron file because in some
systems the script doesn't run with 664 permissions.
It also removes the redundancy of the systemd preset file name and
specify the version that the file enables.
Lastly, now the ubuntu packages don't fail in the post install if
the systemd service doesn't start by an unexpected behaviour like
what happens on docker.
This fixes #145 and #132.

Signed-off-by: Raphael Moreira Zinsly <rzinsly@linux.vnet.ibm.com>