Patch glibc for gcc enabling -mfloat128 by default on AT 11.0
GCC enabled -mfloat128 by default [1].  This requires patching in glibc,
because some files were built on the assumption that compiler support
for float128 had to be explicitly enable in the command line.

The patches to address this issue were accepted on libc-alpha [2], so
this commit applies the patches to AT 11.0.  When the patches get
integrated to the 2.26 branch, this commit will need to be reverted.

[1] https://gcc.gnu.org/ml/gcc-patches/2017-08/msg01028.html

[2] https://sourceware.org/ml/libc-alpha/2017-09/msg00427.html

Signed-off-by: Raphael Moreira Zinsly <rzinsly@linux.vnet.ibm.com>