Update python on AT Next
The test suite in python is composed of many test cases, among them
test_ssl, which itself is composed of many subtests.  AT Next fails in
two of these subtests: test_multiprocessing_forkserver and
test_random_fork, as describe in issue #201.

This commit updates python to version 3.6.3, which fixes the failure
with test_multiprocessing_forkserver.  It also removes a local patch
that was integrated upstream.

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.eti.br>