Fix python test failure (test_random_fork) on AT Next
The test suite in python is composed of many test cases, among them
test_ssl, which itself is composed of many subtests.  AT Next fails in
one of these subtests, test_random_fork, as describe in issue #201.

This commits applies an adapted version of an upstream patch [1], which
fixes the failure with test_random_fork.

Tested for SLES 12, Ubuntu 14.04, and Ubuntu 16.04.

[1] https://hg.python.org/cpython/raw-rev/8e1194c39bed

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.eti.br>