Enforce TLS availability during gcc's stage_optimized
There are some configure scripts executed by GCC that check for the
availability of thread-local storage.  This test requires to build and
run a program.  This is usually safe, except for tuned libraries because
they're built with a -mcpu that may be higher than the cpu where the
build occurs.
This patch enforces TLS availability In order to avoid these tests from
silently failing and reconfiguring GCC's libraries to not use TLS.

This patch fixes issue #204 for libstdc++, libgomp and libitm on AT 10.0
and 11.0.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.vnet.ibm.com>