Enforce vsnprintf availability during gcc's stage_optimized
Libssp has a test to check if vsnprintf is available that requires
running a program.  A cpu-optimized build may have new instructions
that aren't available in the underlying processor, causing this
configure test to silent fail.
So, it's necessary to force the availability of vsnprintf to get the
correct result and avoid this illegal instruction.

This is the final patch fixing issue #204 on AT 10.0 and 11.0.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.vnet.ibm.com>