package upgrade will remove new environment modulefiles
The final step in a package upgrade is %postun.  This is after
%post for the installation of the new version of the package.
The current %postun will unconditionally remove the files/symlinks
from the system modulefiles directory, thus removing the newly
installed environment modulefiles.  The fix is to move the removal
inside the conditional in %postun, which checks to see if the
current uninstall operation is still leaving an installed
instance of the package on the system.

Fixes #365

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>