Use string operator in string comparison
The comparison operator for strings in bash is '=' or '==', whereas the
'-eq' operator is supposed to be used for comparison of integers [1].

For instance, the following code block:

  if [[ c550190e067a -eq fc7e1fcb28eb ]]; \
  then                                    \
    echo equal;                           \
  fi

will print "equal", even though the strings are different.

This patch replaces '-eq' with '==', in update_revision.sh, for the
comparison of revisions, which are usually strings.  If the revision
could be interpreted as an integer (e.g.: svn revisions), interpreting
them as strings still produces the expected result.

[1] https://www.gnu.org/software/bash/manual/bashref.html#Bash-Conditional-Expressions

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.eti.br>