Limit the amount of jobs generating locales in parallel
Locale generation is a high-processing task that uses a lot of RAM/process.
The RAM usage increased recently, causing processes to die in some
servers that had more computing power than RAM.

This unbalanced setup requires to add a restriction to the amount of
jobs running in parallel in order to guarantee the limits will not be
passed.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.vnet.ibm.com>