Downgrade OpenSSL to a stable version on AT next
OpenSSL is a dependency for other packages, specially Python.
The relationship between both packages has been known to cause an extra
burden, i.e. we have a list of Python tests that are ignored
specifically because they're known to fail with certain OpenSSL revisions.
The extra burden increased when AT next started to track OpenSSL's
master branch and its frequent changes.  These frequent modifications
may cause issues that are not instantly fixed for obvious reasons.

This patch is intended to stop the Whac-a-mole behavior and downgrade
OpenSSL to a stable branch on AT next.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>