Disable test_signal and test_epoll in Python's FVTR
Every week, the continuous integration bot (Cherbot) submits around 10
pull requests with updates for many components of the advance-toolchain.
Each pull request is automatically tested, then manually reviewed and
merged by a couple of developers.  Although the automatic creation of
pull requests and testing already reduce the maintenance effort, further
reducing that effort is desirable.

Python's test suite is composed of more than 300 test cases, some of
which fail intermittently.  When that happens, someone typically restart
the build and waits for the results before finishing the review of the
pull request.

This patch disables two of these test cases (test_signal and
test_epoll), with the goal of reducing the maintenance burden, at the
cost of reduced test coverage.

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.eti.br>