Switch OpenSSL to the newest stable branch on AT Next
AT Next does not follow the master branch of OpenSSL, anymore (because
doing so used to cause many failures in the continuous integration),
instead, it follows the latest stable branch.

Since OpenSSL's commit

    commit 1708e3e85b4a86bae26860aa5d2913fc8eff6086
    Author: Matt Caswell <matt@openssl.org>
    Date:   Tue Sep 11 13:48:18 2018 +0100

        Prepare for 1.1.1 release

a new stable branch is available (for OpenSSL 1.1.1).  This patch
updates AT Next to follow this new stable branch.

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.eti.br>