Switch Python to the newest stable branch on AT Next
Upstream python provides a stable branch for the 3.7 release, which AT
Next could follow.  This patch switches to that stable branch.

This patch also adds a patch to disable the checking of the version of
zlib in test_zlib.  This is required because the version of zlib that AT
Next distributes contains a non-numeric suffix.

Finally, this patch adds libffi-devel as a new requirement for AT Next
builds.  If this package is not available in the build machine, the
_cpython Module doesn't get build.

Fixes #576.

Signed-off-by: Gabriel F. T. Gomes <gabriel@inconstante.eti.br>