Update valgrind on AT next
Bump to revision de7fc1a0593c

In addition it takes a newer itrace tarball including an update to fit
with a change in Makefile.am in Valgrind.

In addtion it updates the test of valgrind.
DHAT is no longer experimental since commit 441bfc5[1] (valgrind 3.14.0).
This means the test of valgrind after building AT next needs to load the
library vgpreload_dhat instead of vgpreload_exp-dhat.

[1] https://sourceware.org/git/?p=valgrind.git;a=commit;h=441bfc5f51c7f5f80cc6491d23cbe2dc711d191f

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>