Fix screen detection
In a previous commit, a few lines were added to `remote-build.exp` to
check if screen is installed or not. If not, the script will use tmux
instead of screen.

The lines were misplaced, it checked on the local machine instead of the
build machine. This patch moves the check to the build machine.

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>