Fix cross build of glibc
When building glibc, the systemtap headers are copied to a directory where gcc
looks at. After the build the systemtap headers are removed.
When cross building glibc, 64 bits version and 32 version can be built at the
same time. Meaning a build can copy and remove the systemtap headers while the
other build is compiling ; the systemtap headers will be missing to complete
the second build.

To avoid this, the builds can't be done in parallel. The 64 bits build will
occur before the 32 bits build.

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>