Refactor code for enabling systemtap probes in glibc
Replaces the previous code that enabled systemtap probes in glibc with a
new one that copies the header files to the build directory.
This has the advantage of allowing 2 builds (e.g. 32 and 64 bits) to
execute at the same time because each build has its own copy of the
header files.  Its removal will not break the other build.

This is also refactoring the FVTR testcase in order to improve the log
output and to guarantee the test will check both 32 and 64 bit libraries
when necessary.

Related to issue #946.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>