Add FVTR tests for pveclib
The headers installed by pveclib provide many functions to be inlined in
user code.  This test checks that the headers are installed in the right
place, i.e.: under a pveclib/ sub-directory.  In pre-release versions of
pveclib, the headers were installed directly under the include
directory, which required users to pass `-I<path>' to the compiler, but
this is no longer required, nor desired.

The test source, test.c, is copied directly from pveclib test suite.

Signed-off-by: Gabriel F. T. Gomes <gabrielftg@linux.ibm.com>