Fix: Mixing AT and host env during gcc build stages
On gcc_2 (and similarly with gcc_3), it should use gcc_1, binutils_1 and
glibc_1 but it was using host gcc and because of that mixing at and host
includes and libs.  Also it was mixing running binaries compiled against
AT with runtime from host, using host loader and host libs (instead of
at versions).

This bug does not always present itself if the host and AT libs versions
are sufficiently compatible but presented conflicts in some
combinations, notably AT 11/12 with Debian 9 and (only changing CC/CXX)
AT 13 and Ubuntu 18.04.

By setting CC/CXX it fix all search paths for include and linking while
building gcc_2 and gcc_3.  But It does not fix the library search path
for execution.  This is clear during gcc_2 compilation.  It compiles a
helper program called genautomata and tries to run it as part of the
make script.  It is built against glibc from AT but it is run with the
host loader (/lib64/ld.so.2) which uses the host library paths to search
libs.  In the case of AT13 and ubuntu 18.04 it fails because the host
glibc is too old for the newly build program, and the loader complains.

Also, despite providing and suitable $PATH and gcc configure detecting
powerpc64le-linux-gnu-ar and powerpc64le-linux-gnu-ranlib from AT, in
some places during make it tries to use *-ar and *-ranlib from system.
Which, again, is only apparent when the host dos not provide
powerpc64le-linux-gnu-* on PATH.  Luckily, the build does respect AR and
RANLIB enviroment variables.  So we set them explicitly together with NM
for good measure.  And to make sure the build everything available from
AT toolchain we add it to PATH also on the make step.

Because of CC/CXX changes we can remove some now unused cc_64 and cxx_64

Also reordered all environment variables in configure to a more
consistent order for all stages.

Fix #966 (gcc failed to build on Debian 9)

Signed-off-by: Raoni Fassina Firmino <raoni@linux.ibm.com>