Update GCC source repository
Since January 13th, gcc has moved to git from svn.

Because of the way to 'hide' vendors branches, it needs to update AT < next.
gcc/sources files include git commands to reach IBM branches. For that,
ATSRC_PACKAGE_GIT can be a list of git commands.
It also needs to update the update_revision script to be able to check if
there is an update in IBM branches.

All revisions in gcc/sources files have been updated to match the commit id
in the new gcc repository.

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>