Avoid machine_name fix on GCC's fixincludes step
AT >= 13.0 started to fail on RHEL8 due to a system header fix
applied by GCC on stages >= 2.

- Problem:

When GCC runs the 'fixincludes' step, it searches for libraries that
'need fixing' on the system header directory passed to it during build
configuration. The default value is /usr/include, that is what is
being used on the non-cross build for RHEL8.

During that process, GCC identifies that bits/statx.h needs fixing, and
applies the fixes based on the file from /usr/include. The fixed version
is added to a gcc/include-fixed dir inside GCC's build directory, which
takes precedence when searching for headers over other system header
locations.

This specific header uses a macro called "__glibc_has_include", which is
defined by sys/cdefs.h. However, this macro has been removed from both
files on glibc in November by commit bfa864e1645e.

During the build, GCC is using the headers compiled by glibc on the
previous AT build step, located in
<at-build-dir>/next/opt/at-next-xxx/include, which includes the
sys/cdefs.h header. Since this is from a fairly new glibc revision, the
macro "__glibc_has_include" has already been removed. But the fixed
bits/statx.h header, which is based on an older version installed on the
system under /usr/include, still uses it. This is causing the error, as
the macro is being used with no prior definition.

- Temporary solution:

The breaking fix being applied to bits/statx.h is the 'machine_name'
fix. Based on the comments from {GCC_SRC}/fixincludes/fixinc.in line
130, this fix is not strictly necessary for the build to succeed. So
I disabled it for any powerpc* targets on GCC builds, limited to
AT 13.0 and AT next, where the problem was happening.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>