Force GCC stage 2 bootstrapping to use AT's libc
The build of GCC stage 2 is configured to use --with-advance-toolchain
to guarantee that the linker and headers from AT are used instead of the
system's, which includes setting the -rpath linker flag.

However, this setting does not apply to the tools built during GCC's
bootstrap process. The dynamic linker is currently using the system's
libc.so and AT's ld.so, which can cause problems.

This commit sets the correct -rpath to be used by the linker on stage 1
of the bootstrap to fix this issue.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>