Avoid users branches in update_revision.sh
When checking what was the latest git commit id of a given repository, more
than one branch name could be returned. In this case the given repository is
skipped.
The current patch ignores users branches that could be found (we only want
official branches). Despite this change, if more than one branch is found,
a proper warning will be printed.

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>