Add libnxz to AT next
This commit adds NX GZIP library version v0.60 on AT next toolchain_extra
bundle.

This project uses both GPL 2.0 and Apache 2.0 licences. Therefore the
project was added to the list of GPL projects on the release notes. Also
a copy of the Apache 2.0 license was added to the release notes.

The project source code is also distributed on the advance toolchain
tarball.

Signed-off-by: Lucas A. M. Magalhaes <lamm@linux.ibm.com>