Accommodate hosts with different HOME directories
`remote-build.exp` currently presumes that the user's
local home directory is the same as the user's directory
on the remote host. So, if the local home is, say,
/home/user and the remote home is, say, /home2/user,
and a HOME-relative directory is chosen for the "tray"
or the "destination" (which is the default), the build
will fail.

Accommodate different remote HOME directories by
deferring the determination of dependent variables
until the remote HOME directory can be retreived
and prepended appropriately.

Fixes #1282

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>