Apply zlib optimizations to AT 12, 13 and 14
Commit e6e16619741c introduced patches with optimizations for zlib, but
they were initially only applied to AT next. Since the last zlib release
was in 2017, AT 12 through next all use the same package version, so we
can create a symlink chain and apply the same patch set to all of them.

This also updates the zlib patch to a newer version that includes some
improvements and bug fixes for the optimized functions.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>