Add prefix timestamps to build messages
This is a subjective change, but I believe messages for long-running
tasks should always be timestamped as a best practice. Timestamps can
be diagnotic as well, if a build takes extra time, the culprit phase
can be found quickly.

The chosen format of the timestamp matches that already chosen for
the collected logs.

Sample output:
```
2020-07-30_01.59.31 Completed binutils copy!
2020-07-30_01.59.31 Applying patches to binutils...
2020-07-30_01.59.31 Applied patches to binutils!
2020-07-30_01.59.31 Starting the build of binutils_1...
```

Also, sneaking in two tiny changes:
- Combine two lines which are, together, a single `find` command.
- Spell-o

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>