Use AT's strip when generating deb packages
In the process of generating deb packages for AT, we use dh_strip from
debhelper project to strip our binaries. This script uses the system's
strip, which is almost guaranteed to be an older version than the one we
just built for AT.

Since the AT binaries are generated using a newer binutils, they might
contain symbols the system's strip may not recognize, leading to
problems like:

strip: debian/<long-path>/opt/at-next-15.0-0-alpha/lib64/libc.a(ppc-mcount.o):
	unsupported relocation type 0x74
strip: ppc-mcount.o: bad value

To avoid this, we can force dh_strip to use the strip binary we just
built, which matches the binutils version used when generating the
binaries. Since we already patch dh_strip before using it, we just have
to update that patch and modify it before applying to use the correct AT
bin dir for the current build.

This commit only applies this change to supported deb-based distros
(Ubuntu and Debian), starting from AT 14.0. The way the patch is
applied should guarantee there are no changes to builds for previous
versions/distros. rpm-based distros are also not modified at this
moment, since they use eu-strip instead, which is not provided by AT.

Lastly, we only strip binaries when doing a native build (see
configs/<config>/deb/{monolithic,monolithic_cross}/rules), so the
changes in this commit should not introduce any changes to cross builds.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>