Fix cross compilation mode in GCC optimized libraries
GCC does not recognize automatically configure parameters passed in file
configparms.  It requires to specify that file at configure time using
--cache-file=configparms.

Despite that, GCC target libraries do not reuse the top configure
variables and require to use target_configargs.

This is also fixing a typo in the name of variable cross_compiling.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>