Update glibc and gcc on AT next + a few fixes
In this patch:
    - Bump glibc to revision 4b564f347fdc
    - Bump gcc to revision 37c3c297396af
    - Add -Wno-error=array-parameter to glibc stages
    - Patch GCC to avoid split-stack issue when building libgo

The latest attempts to bump gcc and glibc failed because of a series of
chained errors. It is likely possible to split this commit into a few
others, but it would require carefully tracking which revision
introduced each issue *and* the correct order in which the packages
could be updated to avoid them. Instead, let's do it in one batch.

Bumping glibc alone causes an 'unknown option' error because of a pragma
used to avoid false-positive ‘-Wstringop-overread’ warnings. This type
of warning was added on a more recent gcc revision than we have now.

Also, bumping GCC alone will cause the false-positive mentioned above,
which has not been yet fixed on the current glibc revision we have.

In addition, recent commits on GCC added new warnings and incremented
existing ones to try to catch new kinds of invalid accesses:

    - 3f9a497d1b0 Extend -Warray-bounds to detect out-of-bounds accesses
    to array parameters.
    - baad4c48a85 Extend -Wstringop-overflow to detect out-of-bounds
    accesses to array parameters
    - 6450f07388f Infrastructure & C front end changes for array parameter
    checking (PR c/50584)

These caused many erros in the glibc build. Some of the failures are
under investigation (glibc BZ26647), but more work is still needed.
Recently errors triggered by new warnings like these only triggered
false-positives, so for now we will just add the proper flags to ignore
them.

Finally, a new issue related to gold and libgo appeared when compiling
gcc optimized for P10 (gcc PR97107). A patch is already being discussed
upstream, but we will apply it here off-tree anyways.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>