next: Disable LTO on GCC stage 3
While building our third GCC (gcc_3), the latest link editor (ld) we have is
from our first binutils build, which uses the system's glibc and dynamic linker.
When Link Time Optimization (LTO) is enabled (the default on GCC) and we try to
build libgcc on gcc_3, GCC passes liblto_plugin.so to the linker to be loaded as
a linker plugin. However, this shared object has been built with a more recent
glibc (glibc stage 1) than the system's.

When the linker tries to load this plugin, it may fail if it contains references
to symbols not present in the system's glibc. This happened recently when new
symbols for the stat family of functions were added to glibc 2.33.

To avoid such problems, let's disable LTO on GCC stage 3.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>