Improve script that updates revisions
A check about branch could lead to the master branch whereas the branch
to check isn't related to it.

This `expr index "${branch}" "HEAD"` checks that any chars in "HEAD" is found
in "${branch}". And if so, set the reference branch to "HEAD".
This means if ${branch} is "VALGRIND_3_16_BRANCH", the reference branch will
be "HEAD", which is incorrect.

Updated the check to ensure the whole word "HEAD" is found or not in "${branch}".

Signed-off-by: Erwan Prioul <erwan@linux.ibm.com>