Fix gdb check on AT <= 12.0
Older AT releases supported both ppc and ppc64.  For those releases,
some executables or symlinks were provided in order to simplify usage.
These executables used to have 2 different prefixes:
 - Target-based, e.g. powerpc64-linux-gnu-readelf;
 - Alternative, e.g. powerpc-linux-gnu-readelf.

However, not all executables available as target-based also had an
alternative-based prefix.  gdb is an exception and is only available as
a target-based prefix.

Futhermore, this issue is not reproduced on AT >= 13.0 because these
versions do not support big-endian targets anymore.

This commit is also improving verbosity of the test showing more
executables that were tested.

Fixes: 2781f58f58d545115ab40bddabcb6b5bbf82ae16
Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>