Fix to not skip Python and gdb FVTR tests on AT next
FVTR tests used to be skipped for AT 5 and 6, but the regex to do that was
broken, and matched any AT_MAJOR_VERSION string that contained '5.' or '6.' in
it. For this reason, Python and gdb FVTR tests started being skipped on AT next,
whose AT_MAJOR_VERSION is 15.0. This commit changes the regex to only match at
the beginning of the string, so it only applies to AT 5 and 6.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>