next: Update gdb and use GMP provided by AT on gdb build
Starting with upstream gdb commit 1b4ac058f7, GMP became a required dependency
for building gdb. Some AT builds started to fail because GMP was not installed
in the system and the gdb build was not able to find the GMP provided by AT
either.

Use the new flag `--with-libgmp-prefix` to tell gdb the right place to find the
GMP built by AT so we stop relying on the system's package.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>