Temporarily disable boost on AT next
A recent GCC update is causing boost build to fail. The fix may take a
few days, so meanwhile let's stop building boost so our builds are able
to finish and we can keep bumping the GCC version.

The list of packages to build is derived from the directory names
under <config>/packages/, and a package only gets added if its
corresponding directory has a file named 'sources' in it. Also, by
changing the extension of boost.mk, it does not get included by the
top-level Makefile, so we skip boost's targets.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>