Handle new curl formatting for status codes on package update script
Apparently curl has changed the way it returns the status code, from
adding a header-like line with 'Status: <code> <description>' to a
single line at the top of the output like 'HTTP/<version> <code>'.

This commit makes the necessary adjustments to update_revision.sh to
handle this transition, and should work with older versions of curl as
well.