Fix next to work without RUNPATH on binaries
dh_shlibdeps was working because it uses the RUNPATH to find out the
libraries. We plan to remove RUNPATH so the path must be explicitly
set with -l.

find_dependencies.sh used RUNPATH to find out which ldd to use. Now
it uses the INTERP field that should always be present.

Signed-off-by: Lucas A. M. Magalhaes <lamm@linux.ibm.com>