Use AT's ld.so during GCC stage 3 build
After the removal of GCC stage 2, we started to compile GCC stage 3 (gcc_3) with
mixed software, part coming from the system and part from AT.

The compiler used to build gcc_3 is gcc_1. Since there is no AT glibc available
at the time gcc_1 is built, we cannot use --with-advance-toolchain to force
binaries compiled with it to use AT's loader, so they end up using the system's
loader (and consequently the system's libraries at load time).

But at the time gcc_3 is built, glibc_1 is already available, and its binaries
end up being used during linking, but the executables still have PT_INTERP set
to the system's loader.

So in practice, there is a mismatch between the glibc used during linking, which
comes from AT, and the one found by ld.so at load time, which comes from the
system.

This started causing problems with recent glibc, which added a new
__libc_start_main@@GLIBC_2.34 symbol that is linked into the binaries for gcc_3,
but can't be found by the (system's) dynamic loader at load time, causing
problems as soon as the configure step for gcc_3 starts.

To avoid this mixed usage when building gcc_3, we can force the usage of the
dynamic loader we already have available from glibc_1, which will find and use
AT's glibc instead of the system's.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>