FVTR perf: Append process ID to perf probe event
The perf FVTR test tried to set a probe point using an existing name.
The probe point used by the test is fixed, so if a previous invocation
of the same test was interrupted, the probe point could remain causing
future invocations to fail.

Use a name for the probe point which is much less likely to encounter
such a conflict by appending the process ID.

Fixes #2047

Signed-off-by: Paul A. Clarke <pc@us.ibm.com>