Use git commit -s to sign auto-update commits
Use the '-s' parameter to include the 'Signed-off-by' line to auto-update
commits. If the user also specifies a GITHUB_SIGNATURE as the 5th argument to
update_revision.sh, the final commit will contain two separate signatures, the
one passed as argument and another for the Git user being used to create the
commit.

This allows us to create a commit that is both signed by our CI bot (Cherbot)
and another user, enabling auto-update PRs to properly pass the DCO check.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>